apply selected templates to subscription
